更新合并主版本代码

+++ modules/library/main/src/main/java/org/geotools/renderer/crs/GeographicHandlerFactory.java
@@ -25,11 +25,7 @@ import org.opengis.referencing.crs.CoordinateReferenceSystem;
 import org.opengis.referencing.crs.GeographicCRS;
 import org.opengis.referencing.operation.MathTransform;
 
-/**
- * Returns a {@link ProjectionHandler} for any {@link GeographicCRS}
- *
- * @source $URL$
- */
+/** Returns a {@link ProjectionHandler} for any {@link GeographicCRS} */
 public class GeographicHandlerFactory implements ProjectionHandlerFactory {
 
     static final double MAX_LATITUDE = 89.99;
@@ -63,8 +59,8 @@ public class GeographicHandlerFactory implements ProjectionHandlerFactory {
                     } else {
                         validArea =
                                 new ReferencedEnvelope(
-                                        Float.MAX_VALUE,
-                                        -Float.MAX_VALUE,
+                                        Integer.MAX_VALUE,
+                                        -Integer.MAX_VALUE,
                                         MIN_LATITUDE,
                                         MAX_LATITUDE,
                                         horizontalSourceCrs);
